% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brak_konia.R
\name{brak_konia}
\alias{brak_konia}
\title{Strategia gdy nie ma konia}
\usage{
brak_konia(stado_tmp3, wartosc_w_krolikach_tmp3)
}
\arguments{
\item{stado_tmp3}{Przechowuje wartosc naszego stada}

\item{wartosc_w_krolikach_tmp3}{Przechowuje ile kazdy garunek jest warty w krolikach}
}
\value{
ruch_wymiana to wektor zwracajacy jakie zwierze ma byc wymienione na jakie
}
\description{
Funkcja jest uzywana gdy nie mamy konia, i opracowuje strategie jakie zwierze powinno byc zamienione na jakie.
}

