% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brak_lub_1_kon.R
\name{brak_lub_1_kon}
\alias{brak_lub_1_kon}
\title{Wykonanie strategii gdy mamy mniej niz 2 konie}
\usage{
brak_lub_1_kon(stado_tmp7, wartosc_w_krolikach_tmp7)
}
\arguments{
\item{stado_tmp7}{Przechowuje wartosc naszego stada}

\item{wartosc_w_krolikach_tmp7}{Przechowuje ile kazdy garunek jest warty w krolikach}
}
\value{
zwraca stado po wymianach
}
\description{
Funkcja uzywajac innych funkcji najpierw opracowuje optymalna strategie a nastepnie wykonuje ja zwracajac nam stado po wymianie
}

