% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czy_stac_nas.R
\name{czy_stac_nas}
\alias{czy_stac_nas}
\title{Sprawdzanie czy gracza stac}
\usage{
czy_stac_nas(nr_zwierzecia, stado_tmp, ceny)
}
\arguments{
\item{nr_zwierzecia}{Nr zwierzecia ktore chcemy kupic}

\item{stado_tmp}{Tymczasowy wektor stada gracza}

\item{ceny}{Wektor cen zwierzat}
}
\value{
Zwraca wektor stada po wymianie
}
\description{
Sprawdza czy gracza stac na dane zwierze na podstawie wektora stada i wektora cen
}

