% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czy_wygrana.R
\name{czy_wygrana}
\alias{czy_wygrana}
\title{Czy wygrana}
\usage{
czy_wygrana(stado_tmp)
}
\arguments{
\item{stado_tmp}{tymczasowy (tmp=temporary) wektor stada}
}
\value{
Zwraca wartosc 1 jezeli spelnione sa warunki wygranej
}
\description{
Konczy gre jezeli gracz wygrywa
}

