% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra.R
\name{gra}
\alias{gra}
\title{Pojedyncza gra}
\usage{
gra(strategia, wektor_cen = c(krolik = 1, owca = 6, swinia = 12, krowa = 36,
  kon = 72, maly_pies = 6, duzy_pies = 36))
}
\arguments{
\item{strategia}{Zadana strategia zgodnie z ktora przeprowadzimy gre.}

\item{wektor_cen}{Przechowuje wartosci zwierzat przeliczone na kroliki}
}
\value{
Zwracamy dlugosc trwania gry (liczbe tur).
}
\description{
Funkcja przeprowadza pojedyncza gre zgodnie z zadana strategia
}

