% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeden_kon.R
\name{jeden_kon}
\alias{jeden_kon}
\title{Strategia gdy mamy jednego konia}
\usage{
jeden_kon(stado_tmp6, wartosc_w_krolikach_tmp6)
}
\arguments{
\item{stado_tmp6}{Przechowuje wartosc naszego stada}

\item{wartosc_w_krolikach_tmp6}{Przechowuje ile kazdy garunek jest warty w krolikach}
}
\value{
ruch_zamiana to wektor zwracajacy jakie zwierze ma byc wymienione na jakie
}
\description{
Funkcja jest uzywana gdy mamy dokladnie jednego konia, i opracowuje strategie jakie zwierze powinno byc zamienione na jakie.
}

