% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeden_kon_mniej_niz_127.R
\name{jeden_kon_mniej_niz_127}
\alias{jeden_kon_mniej_niz_127}
\title{Strategia gdy mamy konia i malo innych zwierzat}
\usage{
jeden_kon_mniej_niz_127(stado_tmp5)
}
\arguments{
\item{stado_tmp5}{Przechowuje wartosc naszego stada}
}
\value{
ruch_zamiana zwraca jakie zwierze ma byc wymienione na jakie
}
\description{
Funkcje uruchamiamy gdy mamy jednego konia i zbyt malo zwierzat aby wygrac. Funkcja opracowuje najlepsza dla nas zamiane
}

