% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeden_kon_wiecej_niz_127.R
\name{jeden_kon_wiecej_niz_127}
\alias{jeden_kon_wiecej_niz_127}
\title{Strategia gdy mamy konia i duzo innych zwierzat}
\usage{
jeden_kon_wiecej_niz_127(stado_tmp4)
}
\arguments{
\item{stado_tmp4}{Przechowuje wartosc naszego stada}
}
\value{
ruch_zamana zwraca jakie zwierze ma byc wymienione na jakie
}
\description{
Funkcje uruchamiamy gdy mamy jednego konia i na tyle duzo innych zwierzat ze mozliwa jest juz wygrana. Funkcja opracowuje najlepsza dla nas zamiane
}

