% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddaj_zwierze.R
\name{oddaj_zwierze}
\alias{oddaj_zwierze}
\title{Oddawanie zwierzat ze stada gracza}
\usage{
oddaj_zwierze(nr_zwierzecia, liczba, stado_tmpB)
}
\arguments{
\item{nr_zwierzecia}{Indeks zwierzecia, ktore chcemy oddac}

\item{liczba}{Liczba zwierzat danego gatunku, ktore chcemy oddac}

\item{stado_tmpB}{Tymczasowy wektor stada gracza}
}
\value{
Zwraca wektor stada po zmianie
}
\description{
Oddaje zwierzeta wybranego rodzaju ze stada gracza
}

