% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmnazanie.R
\name{rozmnazanie}
\alias{rozmnazanie}
\title{Rozmnazanie zwierzat}
\usage{
rozmnazanie(stado_tmp12, kostki)
}
\arguments{
\item{stado_tmp12}{Przechowuje aktualne stado na potrzeby funkcji.}

\item{kostki}{Przechowuje wynik rzutu kostkami.}
}
\value{
Aktualny stan stada.
}
\description{
Funkcja przelicza stado rozmnazajace sie zgodnie z rzutem.
}

