% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzut_kostkami.R
\name{rzut_kostkami}
\alias{rzut_kostkami}
\title{Rzut kostkami}
\usage{
rzut_kostkami()
}
\value{
Zwracany jest wynik losowania w postaci wektora z wyrzuconymi zwierzetami.
}
\description{
Funkcja wykonuje losowanie kostkami i zwraca wektor z wynikiem rzutu.
}

