% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_ADA.R
\name{strategia_ADA}
\alias{strategia_ADA}
\title{Strategia gry superfarmer}
\usage{
strategia_ADA(stado_tmp, ceny = c(krolik = 1, owca = 6, swinia = 12, krowa =
  36, kon = 72, maly_pies = 6, duzy_pies = 36))
}
\arguments{
\item{stado_tmp}{Tymczasowy wektor stada gracza}

\item{ceny}{Przechowuje wartosci zwierzat przeliczone na kroliki}
}
\value{
Zwraca stado po wymianach.
}
\description{
Funkcja realizuje strategie gracza
}

