% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wartosc_stada.R
\name{wartosc_stada}
\alias{wartosc_stada}
\title{Przeliczanie stada na kroliki}
\usage{
wartosc_stada(stado_tmp, wartosc_w_krolikach_tmp)
}
\arguments{
\item{stado_tmp}{Przechowuje wartosc naszego stada}

\item{wartosc_w_krolikach_tmp}{Przechowuje ile kazdy garunek jest warty w krolikach}
}
\value{
Zwraca wartosc stada w krolikach
}
\description{
Funkcja przelicza ile nasze stado jest warte w krolikach
}

