% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiecej_niz_1_kon.R
\name{wiecej_niz_1_kon}
\alias{wiecej_niz_1_kon}
\title{Strategia dla wiecej nie jednego konia}
\usage{
wiecej_niz_1_kon(stado_tmp8)
}
\arguments{
\item{stado_tmp8}{Przechowuje wartosc naszego stada}
}
\value{
zwraca stado po wymianach
}
\description{
Funkcja dokonuje wymiany jednego z koni w taki sposob, ze wygrywamy gre
}

