% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilk_i_lis.R
\name{wilk_i_lis}
\alias{wilk_i_lis}
\title{Kontrola wilka i lisa}
\usage{
wilk_i_lis(stado_tmp11, kostki)
}
\arguments{
\item{stado_tmp11}{Przechowuje aktualne stado na potrzeby funkcji.}

\item{kostki}{Przechowuje wynik rzutu kostka}
}
\value{
Zwraca uaktualnione stado.
}
\description{
Funkcja po otrzymaniu wyniku rzutu i aktualnego stada oblicza stan stada w wypadku gdy jego czesc przepada.
}

