% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wymiana_na_drozsze.R
\name{wymiana_na_drozsze}
\alias{wymiana_na_drozsze}
\title{Wymiana zwierzat na drozsze}
\usage{
wymiana_na_drozsze(stado_tmpD, ceny)
}
\arguments{
\item{stado_tmpD}{Tymczasowy wektor stada}

\item{ceny}{Wektor cen zwierzat}
}
\value{
Zwraca wektor stada po wymianie
}
\description{
Kupuje najdrozsze zwierze, na ktore stac gracza
}

