% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wymiana_na_tansze.R
\name{wymiana_na_tansze}
\alias{wymiana_na_tansze}
\title{Wymiana zwierzaka na mniejszego}
\usage{
wymiana_na_tansze(stado_tmpC)
}
\arguments{
\item{stado_tmpC}{Tymczasowy wektor stada gracza}
}
\value{
Zwraca wektor stada po wymianie
}
\description{
Funkcja rozmienia na mniejsze największe zwierze, spośród tych, których gracz ma wiecej niz jedno
}

