% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zamien_na_duze.R
\name{zamien_na_duze}
\alias{zamien_na_duze}
\title{Wymiana malych zwierzat na duze}
\usage{
zamien_na_duze(zwierz, stado_tmp2, wartosc_w_krolikach_tmp2)
}
\arguments{
\item{zwierz}{przechowuje zwierze (konia, swinie lub krowe) na ktore bedziemy chcieli wymienic wiele malych zwierzat}

\item{stado_tmp2}{Przechowuje wartosc naszego stada}

\item{wartosc_w_krolikach_tmp2}{Przechowuje ile kazdy garunek jest warty w krolikach}
}
\value{
Zwraca wektor stada po wymianie
}
\description{
Funkcja otrzymujac informacje na jakie duze zwierze chcemy wymienic male zwierzaki dokonuje odpowiedniej wymiany
}

