% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zamien_zwierze_na_zwierze.R
\name{zamien_zwierze_na_zwierze}
\alias{zamien_zwierze_na_zwierze}
\title{Wymiana zwierzaka na zwierzaka}
\usage{
zamien_zwierze_na_zwierze(oddajemy, dostajemy, stado_tmp1,
  wartosc_w_krolikach_tmp1)
}
\arguments{
\item{oddajemy}{Przechowuje zwierze ktore chcemy wymienic}

\item{dostajemy}{Przechowuje zwierze ktore chcemy otrzymac}

\item{stado_tmp1}{Przechowuje wartosc naszego stada}

\item{wartosc_w_krolikach_tmp1}{Przechowuje ile kazdy garunek jest warty w krolikach}
}
\value{
Zwraca wektor stada po wymianie
}
\description{
Funkcja otrzymujac informacje jakie zwierze ma wymienic na jakie dokonuje tej wymiany
}

