% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{badaj_gre}
\alias{badaj_gre}
\title{Badanie gry}
\usage{
badaj_gre(strategia)
}
\arguments{
\item{strategia}{Przechowuje funkcje strategii gry ktora bedziemy badac.}
}
\description{
Funkcja przeprowadza gre 10.000 razy oraz wylicza statystyki czasu gry.
}

