% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra.R
\name{gra}
\alias{gra}
\title{Pojedyncza gra}
\usage{
gra(strategia)
}
\arguments{
\item{strategia}{Zadana strategia zgodnie z ktora przeprowadzimy gre.}
}
\value{
Zwracamy dlugosc trwania gry (liczbe tur).
}
\description{
Funkcja przeprowadza pojedyncza gre zgodnie z zadana strategia
}

