% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_postMDiPR.R
\name{strategia_postMDiPR}
\alias{strategia_postMDiPR}
\title{Strategia gry superfarmer}
\usage{
strategia_postMDiPR(stado_tmp9)
}
\arguments{
\item{stado_tmp9}{Przechowuje wartosc naszego stada.}
}
\value{
Zwraca stado po wymianach.
}
\description{
Funkcja wykonujaca cala strategie gry superfarmer
}

